HACK: dwc3: Avoid sleep while in atomic in dwc3_gadget_ep_dequeue()
The dwc3 driver has a bug where it calls wait_event_lock_irq()
in the dwc3_gadget_ep_dequeue() function. This drops the specified
lock and sleeps waiting for a condition to become true.

Unfortunately, there isn't only one lock being held, instead the
function is called in irq context, and we already hold an lock_irq
before we enter dwc3_gadget_ep_dequeue(), so when we release the
dwc->lock, and sleep, we are still in atomic context.

This issue has been properly fixed upstream around ~v4.20, and
the fixes were backported to 4.19, however too much has changed
for those fixes to be backported all the way to 4.14.

As this only seems to trip on system shutdown, this hack just
simply removes the conditional wait. With this, I've not seen
any of the crashes on restart that we normally see with 4.14

Signed-off-by: John Stultz <john.stultz@linaro.org>
Change-Id: If3e2ab9fe6a73eb9282d375b3b1135f626794649