ANDROID: hikey960_defconfig: Drop CONFIG_RT_GROUP_SCHED
Recently YongQin reported seeing bluetooth failures on HiKey960
with the 4.14 and 4.19 kernels as tracked in this bug:
  https://bugs.96boards.org/show_bug.cgi?id=889

I narrowed it down to the change updating HiKey960 to FCM6 and
_API_LEVEL 31:
  https://android-review.googlesource.com/c/device/linaro/hikey/+/1759050

It seemed to be connected to the chunk dropping the
schedulerservice here, as we were seeing the sched_setscheduler
calls fail:
  https://android-review.googlesource.com/c/device/linaro/hikey/+/1759050/1/compatibility_matrix.xml

But it didn't make much sense as 5.4/5.10 kernels were not affected.

After a bunch of digging, I found the failure in the kernel was
happening here:
  https://android.googlesource.com/kernel/common/+/refs/heads/android-4.19-stable/kernel/sched/core.c#4934

Which I noticed was in a CONFIG_RT_GROUP_SCHED block, which is *not*
enabled in the 5.4/5.10 gki_defconfig!

This patch removes CONFIG_RT_GROUP_SCHED from the hikey960_defconfig
and gets bluetooth back to working

Reported-by: YongQin Liu <yongqin.liu@linaro.org>
Signed-off-by: John Stultz <john.stultz@linaro.org>
Change-Id: I12b9bc589ca2040946301ace0e428bf998e1441f